+++ CHANGELOG.md
@@ -4,8 +4,9 @@
 
 ### Added
 
-- Support for Fluid Field Groups introduced in ExpressionEngine 7.3
+- Support for Fluid Field Groups and chainable modifiers introduced in ExpressionEngine 7.3
 - Configuration flag to signal preference for using GraphQL Union Types
+
 ### Fixed
 
 - Behavior of `exp.redirect()` tag